Fixes And Delta Of rewards flow for calculating accumulated rewards

+++ src/msg.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 pub struct InstantiateMsg {
     pub vault_denom: String,
     pub amount_to_stake_per_validator: Uint128,
-    pub batch_size: u32,
+    pub batch_size: u64,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -14,7 +14,7 @@ pub struct InstantiateMsg {
 pub enum ExecuteMsg {
     RecordMetrics { timestamp: u64 }, // 12:00AM
     AddValidator { addr: Addr },
-    UpdateConfig { batch_size: u32 },
+    UpdateConfig { batch_size: u64 },
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -35,16 +35,16 @@ pub enum QueryMsg {
     },
     GetValidatorsMetricsByTimestamp {
         timestamp: u64,
-        from: u32,
-        to: u32,
+        from: u64,
+        to: u64,
     },
     GetState {},
     GetConfig {},
     GetAllAprsByInterval {
         timestamp1: u64,
         timestamp2: u64,
-        from: u32,
-        to: u32,
+        from: u64,
+        to: u64,
     },
     GetAprByValidator {
         timestamp1: u64,
